# Conflicts:
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/mapper/EventMapper.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/mapper/ShowMapper.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/entity/Show.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/repository/ShowRepository.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/EventService.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/ShowService.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/EventServiceImpl.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/ShowServiceImpl.java
#	frontend/src/app/app.module.ts

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/interfaces/TicketPrintsApi.java
@@ -68,7 +68,7 @@ public interface TicketPrintsApi {
         produces = { "application/pdf" }
     )
     default ResponseEntity<org.springframework.core.io.Resource> ticketPrintsIdGet(
-        @Parameter(name = "id", description = "ID of the ticket that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Integer id
+        @Parameter(name = "id", description = "ID of the ticket that is retreived", required = true, schema = @Schema(description = "")) @PathVariable("id") Long id
     ) {
         return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
